requirements.txt
telethon>=1.36.0
cryptg>=0.4.0
customtkinter>=5.2.0
PySocks>=1.7.0
pyinstaller>=6.0.0
